+++ packages/terra-section-header/tests/jest/SectionHeader.test.jsx
@@ -4,7 +4,7 @@ import ThemeContextProvider from 'terra-theme-context/lib/ThemeContextProvider';
 import SectionHeader from '../../src/SectionHeader';
 
 describe('SectionHeader', () => {
-  const defaultRender = <SectionHeader title="foo" />;
+  const defaultRender = <SectionHeader text="foo" />;
 
   // Snapshot Tests
   it('should render a default component', () => {
@@ -13,19 +13,19 @@ describe('SectionHeader', () => {
   });
 
   it("should render without an accordion icon when no 'onClick()' is passed", () => {
-    const sectionHeader = <SectionHeader title="foo" />;
+    const sectionHeader = <SectionHeader text="foo" />;
     const wrapper = render(sectionHeader);
     expect(wrapper).toMatchSnapshot();
   });
 
   it("should render with an accordion icon when an 'onClick()' callback is passed", () => {
-    const sectionHeader = <SectionHeader title="foo" onClick={() => {}} />;
+    const sectionHeader = <SectionHeader text="foo" onClick={() => {}} />;
     const wrapper = render(sectionHeader);
     expect(wrapper).toMatchSnapshot();
   });
 
   it("should render with an accordion icon in the open position when an 'onClick()' callback and 'isOpen' is passed", () => {
-    const sectionHeader = <SectionHeader title="foo" onClick={() => {}} isOpen />;
+    const sectionHeader = <SectionHeader text="foo" onClick={() => {}} isOpen />;
     const wrapper = render(sectionHeader);
     expect(wrapper).toMatchSnapshot();
   });
@@ -33,7 +33,7 @@ describe('SectionHeader', () => {
   it('correctly applies the theme context className', () => {
     const wrapper = mount(
       <ThemeContextProvider theme={{ className: 'orion-fusion-theme' }}>
-        <SectionHeader title="foo" />
+        <SectionHeader text="foo" />
       </ThemeContextProvider>,
     );
     expect(wrapper).toMatchSnapshot();